Separate SvDocstring class into separate module.
