+++ utils/sv_extra_search.py
@@ -25,6 +25,7 @@ import sverchok
 from sverchok.menu import make_node_cats
 from sverchok.node_tree import SvDocstring
 from sverchok.utils import get_node_class_reference
+from sverchok.utils.docstring import SvDocstring
 from sverchok.ui.sv_icons import custom_icon
 from sverchok.utils.sv_default_macros import macros, DefaultMacros
 from nodeitems_utils import _node_categories
